Facebook Sign; code refactoring
