+++ .env
@@ -6,8 +6,9 @@ VUE_APP_API_KEY=40d26954d2e35216b139b80e5f442fef
 VUE_APP_API_LANGUAGE=pt-BR
 
 VUE_APP_FIREBASE_API_KEY=AIzaSyA85b_Jz4kK66SxGDSImedViLqYvPYEcc4
+VUE_APP_FIREBASE_AUTH_DOMAIN=wishflix-d0c08.firebaseapp.com
 VUE_APP_FIREBASE_PROJECT_ID=wishflix-d0c08
-VUE_APP_FIREBASE_AUTH_DOMAIN=wishflix-d0c08-default-rtdb.firebaseio.com
-VUE_APP_FIREBASE_DATABASE_URL=https://wishflix-d0c08-default-rtdb.firebaseio.com
-
-
+VUE_APP_FIREBASE_STORAGE_BUCKET=wishflix-d0c08.appspot.com
+VUE_APP_FIREBASE_MESSAGING_SENDER_ID=68430584295
+VUE_APP_FIREBASE_APP_ID=1:68430584295:web:be02160f5e607f68610542
+VUE_APP_FIREBASE_MEASUREMENT_ID=G-Q63WWRDRGR
