+++ src/locales/en.json
@@ -1,4 +1,6 @@
 {
+  "facebook": "Facebook",
+  "alternative_login": "or login with",
   "add_to_my_list": "Add to my list",
   "already_watched":"Already watched",
   "any_movies_selected": "You don't have any movies here.",
