+++ src/locales/pt_BR.json
@@ -1,4 +1,6 @@
 {
+  "facebook": "Facebook",
+  "alternative_login": "ou faça o login com",
   "add_to_my_list": "Adicionar a minha lista",
   "already_watched":"Já assisti",
   "any_movies_selected": "Você não tem filmes aqui.",
