+++ src/plugins/firebase.js
@@ -10,10 +10,11 @@ const env = process.env
 let config = {
     apiKey: env.VUE_APP_FIREBASE_API_KEY,
     authDomain: env.VUE_APP_FIREBASE_AUTH_DOMAIN,
-    databaseURL: env.VUE_APP_FIREBASE_DATABASE_URL,
     projectId: env.VUE_APP_FIREBASE_PROJECT_ID,
-    storageBucket: '',
-    messagingSenderId: ''
+    // storageBucket: env.VUE_APP_FIREBASE_STORAGE_BUCKET,
+    // messagingSenderId: env.VUE_APP_FIREBASE_MESSAGING_SENDER_ID,
+    // appId: env.VUE_APP_FIREBASE_APP_ID,
+    // measurementId: env.VUE_APP_FIREBASE_MEASUREMENT_ID
 }
 
 Vue.prototype.$firebase = firebase
