+++ src/views/users/SignIn.vue
@@ -32,14 +32,26 @@
         ></b-form-input>
       </b-form-group>
 
-      <b-button type="submit" variant="primary" class="btn btn-lg btn-block">
+      <b-button type="submit" variant="success" class="btn btn-lg btn-block">
         {{ $t("sign_in") }}
       </b-button>
+
+      <router-link :to="{ name: 'signup' }">
+        {{ $t("create_an_account") }}
+      </router-link>
     </b-form>
+
+    <hr />
+    <p>{{ $t("alternative_login") }}</p>
+    <b-button variant="primary" size="sm" @click="signInWithFacebook">
+      {{ $t("facebook") }}
+    </b-button>
   </div>
 </template>
 
 <script>
+import firebase from "@/plugins/firebase";
+
 export default {
   name: "SignIn",
   data() {
@@ -51,6 +63,26 @@ export default {
     };
   },
   methods: {
+    signInWithFacebook() {
+      const provider = new firebase.auth.FacebookAuthProvider();
+      this.$firebase
+        .auth()
+        .signInWithPopup(provider)
+        .then((result) => {
+          var credential = result.credential;
+          var user = result.user.displayName;
+          var accessToken = credential.accessToken;
+
+
+// displayName: "Fausto Carvalho"
+// email: "fausto_fcjr@yahoo.com.br"
+// uid
+          console.log(user)
+          console.log(accessToken)
+        })
+        .catch((error) => this.showErrorMessage(error.code))
+        .finally(() => this.$loading(false));
+    },
     signIn(event) {
       event.preventDefault();
       this.$loading(true);
@@ -58,8 +90,8 @@ export default {
         .auth()
         .signInWithEmailAndPassword(this.form.email, this.form.password)
         .then(() => {
-          const currentUser = this.$firebase.auth().currentUser
-          this.$store.commit('setUser',currentUser)
+          const currentUser = this.$firebase.auth().currentUser;
+          this.$store.commit("setUser", currentUser);
           this.$router.push({ name: "user-profile" });
         })
         .catch((error) => this.showErrorMessage(error.code))
